requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
class-registry==2.1.2
exceptiongroup==1.2.2
filters==1.3.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
python-dotenv